Start on automatic FT too
